requirements.txt
asgiref==3.5.2
Django==4.0.7
djangorestframework==3.14.0
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.4
